Update Registry.register.cs
Added ability to forceRebind of the JVM for reregistering an assembly (see my change in Bridge.cs).  This was to allow me to reregister the Bridge assembly with a different class loader.  I am calling into a war file with multiple class loaders and threads.  This worked for me, however I understand that this is sort of a workaround and could be made better for public consumption.